Fixed some code smell
git-svn-id: https://satscm.esac.esa.int/esasky/esasky-app/trunk@5315 aa7853a0-da02-4334-bf23-07ffea9f4f01

+++ esasky-cl/src/main/java/esac/archive/esasky/cl/web/client/model/entities/CombinedSourceFootprintDrawer.java
@@ -21,7 +21,7 @@ public class CombinedSourceFootprintDrawer implements IShapeDrawer{
 	public static final int MAX_LINEWIDTH = 12;
 	private static final double MAX_SECONDARY_SCALE = 10.0;
 
-	private double ratio = DEFAULT_LINEWIDTH / MAX_LINEWIDTH;
+	private double ratio = DEFAULT_LINEWIDTH / (double) MAX_LINEWIDTH;
 
     private double secondaryScale = 1.0;
     private boolean showAvgPM = false;
